web/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.1.0
requests==2.31.0

# Model dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
peft==0.7.1
pyvi==0.1.1
numpy==1.24.3
tqdm==4.66.1